pyproject.toml
[project]
name = "zenvest-ai-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.122.0",
    "google-generativeai>=0.8.5",
    "passlib>=1.7.4",
    "pydantic>=2.12.4",
    "pytest>=9.0.1",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "python-toon>=0.1.3",
    "supabase>=2.24.0",
    "uvicorn[standard]>=0.38.0",
]
